feat: Add version field to Catalog and Snapshot
This commit adds a CatalogVersion type to both InnerCatalog and
CatalogSnapshot in a version field. This does a few things for us:

- It future proofs us so that if we do want to make a breaking change
  in the future we can just add a version to the CatalogVersion enum
  and write code that can handle different versions of the catalog and
  easily upgrade it for the user
- We use an non_exhaustive enum so that code that does depend on it can
  write code to handle a new version of the Catalog for older versions
  of the code

While no code depends on this right now, we definitely will want this in
the future. This is one of those breaking changes that we need before we
head into Beta where we can't do so anymore.