feat: Add version field to PeristedSnapshot
Continuing our work of creating versioned files before Beta, this commit
adds a version field to PersistedSnapshot. This way we can handle
snapshot files that were persisted before being turned into Parquet
files. This way if a user upgrades and we change the version of the file
so that we can make a breaking change, we can just handle that for the
user automatically.