feat: Add versioning for CompactionDetail
This introduces the CompactionDetailVersion type which wraps a
CompactionDetail. The idea is that when we persist or read a file the
version field will let us choose the correct version of CompactedData
to use. This does mean that if we want to make changes to the
CompactionDetail we will have to make sure we create a new field in the
CompactionDetailVersion and a new struct for it. For now the code has
been updated to handle these changes, but we don't need any branching
code to handle different versions since we only have one version now.

A test was also added to make sure that it deserializes with the proper
layout which has the version field inline with the JSON output.