feat: Add versioning to CompactionSummary
This commit, just like the CompactionDetail one before this, adds
versioning to the CompactionSummary. This way we can make breaking
changes after the Enterprise Beta without causing issues for users who
upgrade to newer versions of InfluxDB 3 Enterprise. In much the same
way we add a CompactionDetailVersion enum to handle adding a version
field to the CompactionDetail that we use at the boundaries of reading
and writing from the Object Store. As a result if we add new versions
later the compiler can warn us where we need to handle those differences
in our code.