chore: cleanup debug logs in catalog and related code
There were a lot of excess DEBUG logs littered around the catalog code
from the catalog refactor that were causing a lot of noise in the log
output.

This reduces that noise by:
* removing log statements that were redundant
* simplifying what is included in log statements to make them less verbose
* changing some DEBUG to TRACE in case that verbosity is needed in some
  cases
* several ">>>" were removed from the beginning of DEBUG logs throughout
  the code

In addition, the severity of the message emitted when there are no
subscribers on catalog broadcast was changed from WARN to INFO to not
raise any alarms.