refactor: split existence check to two method in Repository
The `exists` method on the catalog Repository type was split into one
method that only checks for the existence of `id` (for updation) and one
method that checks for the existence of _both_ the `id` and the `name` of
the resource (for insertion).

Having a single method for both was difficult to reason about and may
have led to strange behaviour when we get to renaming of resources.