feat: Add env vars for TLS options
This commit is a follow up to #26246 and adds env vars to all of the TLS
options in Core/Enterprise. This allows users to specify things like a
CA file so that they don't need to run every single command with it
specified. Same with the cert and key files. This is also nice for
letting users configure environments like a Docker container with an
env var rather than with command line args.

Closes #26253