fix: docker: fetch python deps ahead of copying in full source tree
The current build order:

 1. Copy in full source tree
 2. Fetch python build dependencies
 3. Build

The issue with this is any source tree change causes the python dependencies to
be re-fetched, then the rust components of those deps rebuilt.

Instead, copy in just the .circleci directory, which informs the python build
dependency fetch. After we have the python deps, then copy in the full source
tree and build.