feat: port v3 catalog along with conversions (#26464)
This commit brings in v3 catalog changes that's been made in pro. The
pro PR is https://github.com/influxdata/influxdb_pro/pull/866. There is
no real change to catalog that affects core but it is important that
core and enterprise catalogs versions are in sync so that it's easier to
track changes to catalog across core and enterprise.

Includes:
- catalog log file conversions and v3 log file
- catalog snapshot file conversions and v3 snapshot file
- update deserialization function to return v3 version of log/snapshot
  file after applying conversion
- `CatalogSnapshot` and all the types that it depends on that implement
  `Snapshot` trait are moved to `mod.rs` as only recent version of
  snapshot should implement that trait