build(flux): update flux to v0.197.0 (#26564)
* build(flux): update flux to v0.197.0

* chore: update github.com/apache/arrow-go

The new version of flux uses an updated arrow version. Update the
version used throughout influxdb to match.

* fix: staticcheck

Regenerate the protobuf implementations to avoid deprecation warnings
from staticcheck.

* fix: lint

Use the version of protoc-gen-go from the linked protobuf module. This
is to ensure the generated protobuf files are compatible with the
module.

* fix: lint

* fix: lint