feat(influx_tools): report more than one error type (#26600)
Without this PR, the export-parquet tool would report on type conflict
errors and not name conflict errors in the schema if type conflicts were
encountered first. It stopped checking for validation issues once type
conflicts were found.

This PR changes it so that both type and name schema issues are both
identified and reported in the commands output. Either still fails an
export to parquet; but in --dry-run mode the validation is an useful
tool to check for schemas that will be an issue in parquet of
influxdbv3.

* follows #25297