fix(planCompactions): must return all planned groups to release
Planned compaction groups mark the planned files as "in use" in a plan.
The files need to be released if the compaction for a plan isn't run.
That happens "away" from the acquire sites. planCompactionsInner() was
acquiring files into plans and then not returning the plans so the files
could never be released and therefore never incorporated into a future
plan.