fix: full compactions not scheduled under some circumstances
Fix issue introduced with 1.12.1 that caused some TSM files to not be
scheduled for full compaction when they should be. The user visible
symptom was an unbounded increase in disk space for impacted shards.

This simplifies the compaction planning code by removing the
PT_SmartOptimize and PT_NoOptimize special cases. These are no longer
needed for performance thanks to the first block caching code that was
also added in #26432. Issues with the PT_SmartOptimize and PT_NoOptimize
cases could cause TSM files to be locked by the compaction planner as
in-use, preventing them from being compacted.

The optimized compaction hold-off is now handled entirely in
Engine.compact rather than by the compaction planner.

Update test cases as necessary for changes, plus add checks for issue
that caused #26667.

Closes: #26667