fix(security): remove unsafe signal_handler code (#26685)
This commit addresses the fact that our signal_handler function on Unix
systems is not async signal safe, due to the fact that doing heap
allocation to print out the signal and stack trace as well as getting
the current thread name and grabbing a backtrace are not guaranteed to
be unsafe.

We can figure out a way to print out this information in the future
possibly, but in order to remove the possible vulnerability this commit
simply leaves a comment with context and just sets us to abort the
process.

Closes influxdb_pro/971 which contains the security report that brought
this issue to our attention.