fix(tsm1): fix condition check for optimization of array cursor
The array ascending cursors do not check all the required conditions
before doing an optimization to copy a full 1000 points of data from the
tsm read buffer to the cursors response buffer. This can result in
points from tsm files appearing in cursor output multiple times.

The conditions the bug appears are rare as it requires cache data that
is older (or overlapping) with tsm data. The data must be in the same
shard and must match for the same key.

* fixes #26690