chore(tsm1): add test that covers array cursor bug
This commit adds a test that verifies the duplicate data bug fix. It
will fail without the previous commit. See the test for details; in
short it sets up the right cache and tsm data to expose the bug which
was occuring when the cache had older data than the tsm files and then
the cache was exhausted while there were still more tsm values to
include in more Next() calls.