feat(write): add no_sync option to the write command (#26703)
* feat(write): add no_sync option to the write command

The influxdb3 write command was lacking a way to set the no_sync option
which tells the server to acknowledge the write request before it is
durable. This changeset provides a cli flag to configure that option and
wire it into the client call. It adds a test that makes use of the new
setter. It relies on other tests that the api works as it should.

A few typos and a minor update the precision flag, to include the
`auto` option, were also included.

This will make using the write command for importing line protocol more
flexible.

* closes #26702