chore(test): switch to using the testify library in the new test
The assert package is used for post conditions and the require package
is used for pre-conditions in the test. The library provides clearer
intent of conditions and better formatting of information (message and
data) in the situation an assertion fails. It also makes it easier to
read by reducing ifs but some lines are longer.

This commit addresses PR requests.