chore: switch to setpriv to avoid PAM
We want to run a command as another user but su has a frustrating syntax
for calling a command and escaping, runuser is simpiler but delegates to
su so both use PAM which is not needed in this case. It was recommended
to use setpriv which a full toolkit for setting privilege bits and can
mimic su/runuser by setting privilege to a specific user for running a
command. It seems to work and be easy to use in a script.