fix(tsm1): fix condition check for optimization of array cursor (#26691)
* fix(tsm1): fix condition check for optimization of array cursor

The array ascending cursors do not check all the required conditions
before doing an optimization to copy a full 1000 points of data from the
tsm read buffer to the cursors response buffer. This can result in
points from tsm files appearing in cursor output multiple times.

The conditions the bug appears are rare as it requires cache data that
is older (or overlapping) with tsm data. The data must be in the same
shard and must match for the same key.

* fixes #26690

* chore(tsm1): add test that covers array cursor bug

This commit adds a test that verifies the duplicate data bug fix. It
will fail without the previous commit. See the test for details; in
short it sets up the right cache and tsm data to expose the bug which
was occuring when the cache had older data than the tsm files and then
the cache was exhausted while there were still more tsm values to
include in more Next() calls.

* chore(test): switch to using the testify library in the new test

The assert package is used for post conditions and the require package
is used for pre-conditions in the test. The library provides clearer
intent of conditions and better formatting of information (message and
data) in the situation an assertion fails. It also makes it easier to
read by reducing ifs but some lines are longer.

This commit addresses PR requests.