fix(packaging): init.sh should run buildtsi as influxdb user (#26699)
* fix(packaging): init.sh should run buildtsi as influxdb user

The init.sh script assumes that it being run as root so thus we can use
sudo to switch user to the influxdb user to run buildtsi; otherwise the
files are owned by root and influxdb can't start.

* fixes #26698

* chore: switch from sudo to su; update comment

* chore: switch to runuser as the su syntax and debugging is difficult

* chore: switch to setpriv to avoid PAM

We want to run a command as another user but su has a frustrating syntax
for calling a command and escaping, runuser is simpiler but delegates to
su so both use PAM which is not needed in this case. It was recommended
to use setpriv which a full toolkit for setting privilege bits and can
mimic su/runuser by setting privilege to a specific user for running a
command. It seems to work and be easy to use in a script.

* chore: update comment to reflect setpriv usage