chore(cli): hide the token's env var value in cli help
Use clap crate's "hide_env_values" option to suppress displaying the
current value of the env var of the db token INFLUXDB3_AUTH_TOKEN when
displaying cli help. The default behavior of clap is to display env var
values for arguments when displaying the generated help messages.
Showing the token disrupts formatting of the help because it is often
long and in a way leaks the token's value.

* closes #26713