fix(wal_replay): limit the number of wal files preloaded to num_cpus (#26716)
* fix(wal_replay): limit the number of wal files preloaded to num_cpus

Wal replay currently loads all the wal files into memory and decodes
them by default. If that's 10s of GB or 100s of GB, it'll try to do it
potentially causing OOMs if it exceeds system memory. We likely keep
most of the speed from preloading but decrease chance of OOM by
preloading a more limited number of wal files. In the absence of an
option to directly limit the memory used in preload, we can use the
number of cpu cores available as a proxy. This will be the number of wal
files loaded to replay, which has to happen in order still. The current
recommendation is to use 10 if you encounter an OOM so let's use that as
the minimum if a specific value isn't set. The new logic is

num_files_preloaded = (user's choice) of if not set (max(10, num_cpus)

This should improve the experience restarting the server when there is a
lot of wal data.

* closes #26715

* chore: update the default value in help-all

* fix(wal_replay): implement dynamic default value in clap derive for concurrency limit

- Add wal_replay_concurrency_limit_default() helper fn for max(num_cpus, 10)
- Change field type from Option<usize> to usize
- Update help text to clarify dynamic nature and OOM warning

* chore: typo in help